# gudangku/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GudangKu",
        "description": "API Documentation for GudangKu",
        "contact": {
            "email": "flazen.edu@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/report/delete/item/{id}": {
            "delete": {
                "tags": [
                    "Report"
                ],
                "summary": "Hard delete report item by id",
                "operationId": "c34be0425e5d835998032cc959162f9c",
                "responses": {
                    "200": {
                        "description": "report item deleted"
                    },
                    "404": {
                        "description": "report item failed to deleted"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/report/delete/report/{id}": {
            "delete": {
                "tags": [
                    "Report"
                ],
                "summary": "Hard delete report by id",
                "operationId": "7cdf0bd847708c5176cb9630ca860e54",
                "responses": {
                    "200": {
                        "description": "report deleted"
                    },
                    "404": {
                        "description": "report failed to deleted"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}